requirements.txt
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.21
Pillow==7.2.0
protobuf==3.13.0
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyqt5-tools==5.15.0.1.7.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dotenv==0.14.0
PyTweening==1.0.3
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
